Added files via upload
The Players.java file has been tested and finalized. The Deck.java file was created to make the initial Revel code an object of its own in order to clarify and reduce clutter in the main method. There is the possibility of further updates to the Deck.java file, but this initial push is a good start point.